+++ src/main/kotlin/com/soom/monolithic_api/domain/account/common/data/dto/AccountDto.kt
@@ -1,0 +1,7 @@
+package com.soom.monolithic_api.domain.account.common.data.dto
+
+sealed class AccountDto(
+    val basic: AccountBasicDto,
+    val auth: AccountAuthDto,
+    val meta: AccountMetaDto
+)
