+++ src/main/kotlin/com/soom/monolithic_api/domain/account/common/data/dto/StudentDto.kt
@@ -1,0 +1,6 @@
+package com.soom.monolithic_api.domain.account.common.data.dto
+
+class StudentDto(
+    basic: AccountBasicDto, auth: AccountAuthDto, meta: AccountMetaDto,
+    val student: StudentAdditionalDto
+): AccountDto(basic, auth, meta)
