+++ src/main/kotlin/com/soom/monolithic_api/domain/account/common/data/dto/TeacherDto.kt
@@ -1,0 +1,6 @@
+package com.soom.monolithic_api.domain.account.common.data.dto
+
+class TeacherDto (
+    basic: AccountBasicDto, auth: AccountAuthDto, meta: AccountMetaDto,
+    val teacher: TeacherAdditionalDto
+): AccountDto(basic, auth, meta)
