+++ src/main/kotlin/com/soom/monolithic_api/domain/account/common/data/type/AccountType.kt
@@ -1,0 +1,5 @@
+package com.soom.monolithic_api.domain.account.common.data.type
+
+enum class AccountType {
+    학생, 교사
+}
