+++ src/main/kotlin/com/soom/monolithic_api/domain/account/common/data/type/GenderType.kt
@@ -1,0 +1,5 @@
+package com.soom.monolithic_api.domain.account.common.data.type
+
+enum class GenderType {
+    여성, 남성
+}
