+++ src/main/kotlin/com/soom/monolithic_api/domain/account/common/data/type/RoleType.kt
@@ -1,0 +1,5 @@
+package com.soom.monolithic_api.domain.account.common.data.type
+
+enum class RoleType {
+    //TODO
+}
