+++ src/main/kotlin/com/soom/monolithic_api/domain/account/common/data/type/TeacherType.kt
@@ -1,0 +1,6 @@
+package com.soom.monolithic_api.domain.account.common.data.type
+
+enum class TeacherType {
+    //TODO 교사 분류 세분화 예정 | 2022.04.01
+    학교, 기숙사, 그외
+}
