+++ src/main/kotlin/com/soom/monolithic_api/domain/account/common/repository/AccountRepository.kt
@@ -1,0 +1,6 @@
+package com.soom.monolithic_api.domain.account.common.repository
+
+import com.soom.monolithic_api.domain.account.common.data.entity.AccountEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface AccountRepository: JpaRepository<AccountEntity, Long>
