+++ src/main/kotlin/com/soom/monolithic_api/domain/account/common/repository/StudentRepository.kt
@@ -1,0 +1,7 @@
+package com.soom.monolithic_api.domain.account.common.repository
+
+import com.soom.monolithic_api.domain.account.common.data.entity.StudentEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface StudentRepository : JpaRepository<StudentEntity, Long>
+
