--- src/main/kotlin/com/soom/monolithic_api/domain/account/dto/AccountDto.kt
@@ -1,7 +1,0 @@
-package com.soom.monolithic_api.domain.account.dto
-
-sealed class AccountDto(
-    val basic: AccountBasicDto,
-    val auth: AccountAuthDto,
-    val meta: AccountMetaDto
-)
