--- src/main/kotlin/com/soom/monolithic_api/domain/account/dto/ProfileImageDto.kt
@@ -1,3 +1,0 @@
-package com.soom.monolithic_api.domain.account.dto
-
-data class ProfileImageDto (val userId: Long, val imageId: String)
