--- src/main/kotlin/com/soom/monolithic_api/domain/account/dto/StudentDto.kt
@@ -1,6 +1,0 @@
-package com.soom.monolithic_api.domain.account.dto
-
-class StudentDto(
-    basic: AccountBasicDto, auth: AccountAuthDto, meta: AccountMetaDto,
-    val student: StudentAdditionalDto
-): AccountDto(basic, auth, meta)
