+++ src/main/kotlin/com/soom/monolithic_api/domain/account/profile/dto/ProfileImageDto.kt
@@ -1,0 +1,3 @@
+package com.soom.monolithic_api.domain.account.profile.dto
+
+data class ProfileImageDto (val userId: Long, val imageId: String)
