+++ src/main/kotlin/com/soom/monolithic_api/domain/account/profile/response/ProfileImageResponse.kt
@@ -1,0 +1,19 @@
+package com.soom.monolithic_api.domain.account.profile.response
+
+import com.soom.monolithic_api.domain.account.profile.dto.ProfileImageDto
+
+//프로필 이미지 정보를 담은 응답
+sealed class ProfileImageResponse (
+    val userId: Long,
+    val imageId: String
+)
+class EditProfileImageResponse (userId: Long, imageId: String): ProfileImageResponse(userId, imageId) {
+    companion object {
+        fun of(dto: ProfileImageDto) = EditProfileImageResponse(dto.userId, dto.imageId)
+    }
+}
+class AddProfileImageResponse(userId: Long, imageId: String) : ProfileImageResponse(userId, imageId) {
+    companion object {
+        fun of(dto: ProfileImageDto) = AddProfileImageResponse(dto.userId, dto.imageId)
+    }
+}
