+++ src/main/kotlin/com/soom/monolithic_api/domain/account/profile/service/AccountProfileImageService.kt
@@ -1,0 +1,9 @@
+package com.soom.monolithic_api.domain.account.profile.service
+
+import com.soom.monolithic_api.domain.account.profile.dto.ProfileImageDto
+import org.springframework.web.multipart.MultipartFile
+
+interface AccountProfileImageService {
+    fun save(id: Long, image: MultipartFile): ProfileImageDto
+    fun delete(id: Long)
+}
