--- src/main/kotlin/com/soom/monolithic_api/domain/account/repository/StudentRepository.kt
@@ -1,7 +1,0 @@
-package com.soom.monolithic_api.domain.account.repository
-
-import com.soom.monolithic_api.domain.account.entity.StudentEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface StudentRepository : JpaRepository<StudentEntity, Long>
-
