--- src/main/kotlin/com/soom/monolithic_api/domain/account/repository/TeacherRepository.kt
@@ -1,6 +1,0 @@
-package com.soom.monolithic_api.domain.account.repository
-
-import com.soom.monolithic_api.domain.account.entity.TeacherEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface TeacherRepository : JpaRepository<TeacherEntity, Long>
