--- src/main/kotlin/com/soom/monolithic_api/domain/account/response/ProfileResponse.kt
@@ -1,9 +1,0 @@
-package com.soom.monolithic_api.domain.account.response
-
-import com.soom.monolithic_api.domain.account.type.AccountType
-
-//프로필 식별 정보를 담은 응답
-sealed class ProfileResponse (
-    val id: Long,
-    val type: AccountType
-)
