--- src/main/kotlin/com/soom/monolithic_api/domain/account/service/AccountProfileService.kt
@@ -1,7 +1,0 @@
-package com.soom.monolithic_api.domain.account.service
-
-import com.soom.monolithic_api.domain.account.dto.AccountDto
-
-interface AccountProfileService {
-    fun getProfile(id: Long): AccountDto
-}
