--- src/main/kotlin/com/soom/monolithic_api/domain/account/template/AccountTemplate.kt
@@ -1,27 +1,0 @@
-package com.soom.monolithic_api.domain.account.template
-
-import com.soom.monolithic_api.domain.account.entity.AccountEntity
-import com.soom.monolithic_api.domain.account.entity.StudentEntity
-import com.soom.monolithic_api.domain.account.entity.TeacherEntity
-
-interface AccountTemplate {
-    fun <T> doWithAccountById(
-        id: Long,
-        doWithAccount:  (AccountEntity) -> T
-    ): T
-
-    fun <T> doAndGetWithAccountById(
-        id: Long,
-        doWithAccount:  (AccountEntity) -> T
-    ): T
-
-    fun doWithAccountById(
-        id: Long,
-        doWithTeacher: (TeacherEntity) -> Unit, doWithStudent: (StudentEntity) -> Unit
-    )
-
-    fun <T> doAndGetWithAccountById(
-        id: Long,
-        doWithTeacher: (TeacherEntity) -> T, doWithStudent: (StudentEntity) -> T
-    ): T
-}
