--- src/main/kotlin/com/soom/monolithic_api/domain/account/type/AccountType.kt
@@ -1,5 +1,0 @@
-package com.soom.monolithic_api.domain.account.type
-
-enum class AccountType {
-    학생, 교사
-}
