--- src/main/kotlin/com/soom/monolithic_api/domain/account/type/DepartmentType.kt
@@ -1,7 +1,0 @@
-package com.soom.monolithic_api.domain.account.type
-
-enum class DepartmentType(val school: SchoolType) {
-    소프트웨어_공통과정(SchoolType.광주_소프트웨어_마이스터_고등학교),
-    소프트웨어_개발(SchoolType.광주_소프트웨어_마이스터_고등학교),
-    스마트_IOT(SchoolType.광주_소프트웨어_마이스터_고등학교)
-}
