--- src/main/kotlin/com/soom/monolithic_api/domain/account/type/GenderType.kt
@@ -1,5 +1,0 @@
-package com.soom.monolithic_api.domain.account.type
-
-enum class GenderType {
-    여성, 남성
-}
