--- src/main/kotlin/com/soom/monolithic_api/domain/account/type/RoleType.kt
@@ -1,5 +1,0 @@
-package com.soom.monolithic_api.domain.account.type
-
-enum class RoleType {
-    //TODO
-}
