--- src/main/kotlin/com/soom/monolithic_api/domain/account/type/SchoolType.kt
@@ -1,10 +1,0 @@
-package com.soom.monolithic_api.domain.account.type
-
-enum class SchoolType  (
-    val schoolName: String
-        ) {
-    광주_소프트웨어_마이스터_고등학교("광주소프트웨어마이스터고등학교"),
-    대구_소프트웨어_마이스터_고등학교("대구소프트웨어마이스터고등학교"),
-    대덕_소프트웨어_마이스터_고등학교("대덕소프트웨어마이스터고등학교"),
-    부산_소프트웨어_마이스터_고등학교("부산소프트웨어마이스터고등학교")
-}
