--- src/main/kotlin/com/soom/monolithic_api/domain/account/type/TeacherType.kt
@@ -1,6 +1,0 @@
-package com.soom.monolithic_api.domain.account.type
-
-enum class TeacherType {
-    //TODO 교사 분류 세분화 예정 | 2022.04.01
-    학교, 기숙사, 그외
-}
