+++ src/main/kotlin/com/soom/monolithic_api/global/service/DummyLoginAccountService.kt
@@ -1,10 +1,10 @@
 package com.soom.monolithic_api.global.service
 
-import com.soom.monolithic_api.domain.account.dto.*
-import com.soom.monolithic_api.domain.account.type.AccountType
-import com.soom.monolithic_api.domain.account.type.DepartmentType
-import com.soom.monolithic_api.domain.account.type.GenderType
-import com.soom.monolithic_api.domain.account.type.SchoolType
+import com.soom.monolithic_api.domain.account.common.data.dto.*
+import com.soom.monolithic_api.domain.account.common.data.type.AccountType
+import com.soom.monolithic_api.domain.account.common.data.type.DepartmentType
+import com.soom.monolithic_api.domain.account.common.data.type.GenderType
+import com.soom.monolithic_api.domain.account.common.data.type.SchoolType
 import org.springframework.stereotype.Service
 import java.time.LocalDate
 import java.time.LocalDateTime
