+++ src/main/kotlin/com/soom/monolithic_api/global/service/LoginAccountService.kt
@@ -1,6 +1,6 @@
 package com.soom.monolithic_api.global.service
 
-import com.soom.monolithic_api.domain.account.dto.AccountDto
+import com.soom.monolithic_api.domain.account.common.data.dto.AccountDto
 
 interface LoginAccountService {
     fun getLoginAccount(): AccountDto
